Move some functions to site_content
